refactor(map): extract user-position / contact-geocoding / live-presence (4/7)
Step 4 of the PulseMapView strangler decomposition. Pure move — no
behaviour change. Effect dependencies, error-handling branches, and the
in-flight geocode cache all preserve their prior semantics.

hooks/useUserPosition.ts
  - One-shot getCurrentUserLocation read on mount. Flips geoBlocked when
    the rejection carries GeolocationPositionError.code === 1
    (PERMISSION_DENIED). Returns { userPosition, geoBlocked }.

hooks/useContactGeocoding.ts
  - Side-effect hook. Owns its own geocodeRequestedRef so the caller
    doesn't have to plumb a Set through. Re-runs whenever the contacts
    array reference changes. Returns a resetRequested() callback the
    empty-state "Auto-geocode" CTA uses to clear the in-flight cache.

hooks/useLivePresence.ts
  - subscribeToUserLocation per linked contact; accumulates the live
    positions into a Map<contactId, UserLocation>. Cleans up all
    subscriptions on contacts-array change.

PulseMapView.tsx
  - Three inline blocks (~70 lines) collapse to three hook call sites
    plus the resetContactGeocoding callback the Auto-geocode handler
    invokes.
  - Drops getCurrentUserLocation, geocodeContactsBatch,
    saveContactLocation, subscribeToUserLocation imports — all only
    used inside the extracted hooks. UserLocation stays (the
    liveBroadcasters memo at line ~582 still types its tuple with it).
  - userPosition / geoBlocked / liveLocations state declarations drop
    from the component body; the hooks own them now.
